Initial work on jar-add-local sub-generator
alfresco-module-manager still needs to handle
local jar modules though!
