+++ package.json
@@ -28,7 +28,7 @@
     "ascii-table": "0.0.9",
     "chalk": "2.3.0",
     "debug": "3.1.0",
-    "generator-alfresco-common": "0.9.13",
+    "generator-alfresco-common": "0.9.14",
     "lodash": "4.17.4",
     "mem-fs-editor": "3.0.2",
     "rmdir": "1.2.0",
